requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.6
djangorestframework==3.13.1
Pillow==9.2.0
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uuid==1.30
